+++ frontend/src/scss/components/topProductCard.scss
@@ -1,0 +1,45 @@
+.product-grid {
+    &__card {
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start; 
+      height: 100%;
+      margin-bottom: toRem(40px); 
+      
+      &__img-container {
+        height: toRem(200px);
+        width: 100%;
+        overflow: hidden;
+    
+        img {
+          height: 100%;
+          width: auto;
+          object-fit: cover;
+        }
+      }
+    
+      &__info {
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-start; 
+        height: auto;
+    
+        &__title {
+          font-size: toRem(16px); 
+          line-height: 1.2; 
+          height: auto; 
+          overflow: hidden;
+          text-overflow: ellipsis;
+          display: -webkit-box;
+          -webkit-line-clamp: 2;
+          -webkit-box-orient: vertical;
+          margin-bottom: toRem(8px); 
+        }
+    
+        &__price {
+          font-weight: bold; 
+        }
+      }
+    }
+  }
+
