+++ frontend/src/scss/style.scss
@@ -13,3 +13,4 @@
 // components
 @import 'components/ProductCard.scss';
 @import 'components/CategoryComponent.scss';
+@import 'components/topProductCard.scss';
